Allow users to drag and drop ships during decision phases
Notes:
Currently, this uses a swapping implementation.

It's still buggy in places. There are times when ships will not switch
places.  There are also times when two ships will occupy the same
space.

Additionally, the dragged element is not brought to the front of the
canvas, so it will show up behind other elements as it gets dragged.