Use PriorityQueue as the interface between 0 and 1 indexing
It was getting confusing knowing which indices to use.

This commit also fixes a bug.
Ships inserted with the second lowest priority would actually use the
lowest priority.  When the ship is removed from the queue, the queue's
Count reduces by one, so now priority 2 (of 4, indexed starting at
zero) satisfies the old condition, and the ship gets added to the end
of the queue instead of inserted at the proper place.