Serialze ships and shiplistentries and send over the network
Big commit, apologies.  This is all related to various synchonization
stuff relating to the ships.  There are a few things left unsynced but
it pretty much covers it. It syncs the Ship (with location and cargo
data) and the heading from the shipcontroller.  The ShipListEntry has
to send an RPC to add the entry for the ship, but the entries
themselves are not synced.  They get all their information from the
shipcontroller and priority queue.